# backend/swagger.yaml
openapi: 3.0.0
info:
  title: MSE - Main Solution Estates
  description: MongoDB ExpressJS NodeJS REST API for a Real Estate App.<br />Github repository available [here](https://github.com/BenjaminKing1337/fullstack_project-MSE).<br />Some endpoints require `auth-token` in order to be executed.
  version: 1.0.0
servers:
  - url: https://mse-a9ir.onrender.com/api/
    description: Remote server
  - url: http://localhost:3000/api/
    description: Localhost development server
components:
  securitySchemes:
    ApiKeyAuth: 
      type: apiKey
      in: header 
      name: auth-token
  schemas:
...
    Tenant:
      type: object
      properties:
        type:
          type: string
          enum: ["Owner", "Renter"]
        forename:
          type: string
        surname:
          type: string
        email:
          type: string
        phone_number:
          type: string
        keys_number:
          type: string
        closest_neighbour:
          type: string
        account_number:
          type: string
        move_in:
          type: string
        move_out:
          type: string
        lease:
          type: string
        user_id:
          type: string
        created_by:
          type: string
          description: "Id of the logged in user"
        created_at:
          type: string
          description: "It takes the current date as default"
          example: "2014-01-01T23:28:56.782Z"
...
paths:
...
  /tenants:
    get:
      summary: Get all tenants
      tags:
        - Tenant
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Tenants successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tenants/get/byUser/{id}:
    get:
      summary: Get tenants by logged in user ID
      tags:
        - Tenant
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Tenants successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tenants/get/{id}:
    get:
      summary: Get tenant by ID
      tags:
        - Tenant
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Tenant successfully retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tenants/new:
    post:
      summary: New tenant
      tags:
        - Tenant
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
      responses:
        200:
          description: Tenant successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tenants/update/{id}:
    put:
      summary: Update tenant by ID
      tags:
        - Tenant
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tenant"
      responses:
        200:
          description: Tenant successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /tenants/delete/{id}:
    delete:
      summary: Delete tenant by ID
      tags:
        - Tenant
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        200:
          description: Tenant successfully deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Access Denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        498:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...